Refactor default_recipe.yaml to use 'sampling_size' instead of 'sampling_ratio'
The code changes update the 'default_recipe.yaml' file in both the 'knowledge' and 'skills' directories. The 'sampling_ratio' property has been replaced with 'sampling_size' to provide a more intuitive representation of the dataset sampling. This change ensures consistency and clarity in the configuration files. Also removed a new line from the end of sys_prompt.

Signed-off-by: shiv <shivchander.s30@gmail.com>
Signed-off-by: Ben Browning <bbrownin@redhat.com>