feat: Introduce Recipe class and refactor datamixing
Add Recipe class which defines the datasets which will be included in the mixing and their sampling proportions - this is an extremely useful asset to tag models with and for debugging.

Introduce the hard requirement of allowed columns - id, messages, and metadata - where metadata plays the role of catch-all for all other columns and info.

These changes enhance the efficiency and consistency of the datamixing functionality.

Signed-off-by: shiv <shivchander.s30@gmail.com>
Signed-off-by: Ben Browning <bbrownin@redhat.com>