Remove the placeholder datasets in skills default_recipe.yaml
Eventually there will be a dataset (on HuggingFace or elsewhere) that
we'll want to use with every skill. It's not released yet, so for now
remove the placeholder value that causes issues because it doesn't
reference a real path. We will want a precomputed dataset adding in
here, so this is a temporary hack to move things along but will result
in worse skill training results until fixed.

Signed-off-by: Ben Browning <bbrownin@redhat.com>