Fix skill taxonomy sometimes misclassified as knowledge
If a user has multiple taxonomy changes to both skills and knowledge,
we were accidentally classifying any skill taxonomy leaf nodes as
knowledge if they happened to come after a knowledge leaf node. This
was just a case of the tracking variable `is_knowledge` not being
reset to its default False value every iteration of the loop.

The simple e2e test randomly triggers this issue, as an example,
depending on the order the leaf nodes get picked in.

Signed-off-by: Ben Browning <bbrownin@redhat.com>