Remove unused auxiliary dataset code
None of the automated or manual tests appear to exercise any code
paths that would actually create an auxiliary dataset. And, given that
there was a typo in the path to the `auxilary_instructions.yaml` file
it wouldn't have worked even if some code did hit that path.

If this is needed for some reason, it can be added back later in a
separate PR so it's clear what it does and why it's needed.

Signed-off-by: Ben Browning <bbrownin@redhat.com>