Defer recipe yaml files to a separate PR
This removes the entire concept of loading initial datasets and system
prompts from recipe yaml files. It also removes writing out those
files, thus removing the entire user-visible interface of recipe
files. The reason for this is to separate out those specific concerns
about file format, naming, supportability, and so on to a separate PR
to minimize the core concerns of creating mixed datasets in this PR.

Signed-off-by: Ben Browning <bbrownin@redhat.com>